pyproject.toml
[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[project]
name = "gse146912-pipeline"
version = "0.1.0"
description = "Reproducible scRNA pipeline for GSE146912 glomerulus atlas"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "anndata>=0.10",
  "scanpy>=1.9",
  "pandas>=2.0",
  "numpy>=1.24",
  "scipy>=1.10",
  "pyyaml>=6.0",
  "mygene>=3.2",
  "requests>=2.28",
  "matplotlib>=3.7",
]

[project.optional-dependencies]
full = [
  "liana>=1.5",
  "decoupler>=2.0",
  "gseapy>=1.0",
  "pybiomart",
  "scrublet",
]
dev = ["pytest", "ruff"]

[tool.setuptools.packages.find]
where = ["src"]
